add entidade funcionalidades e uma penca de ajustes

+++ src/projetos/projetos.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, HttpException, HttpStatus } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateProjetoDto } from './dto/create-projeto.dto';
 import { UpdateProjetoDto } from './dto/update-projeto.dto';
@@ -8,25 +8,47 @@ export class ProjetosService {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createProjetoDto: CreateProjetoDto) {
-    const projeto = await this.prisma.projetos.create({
-      data: createProjetoDto,
+    const projeto = await this.prisma.projetos.findFirst({
+      where: {
+        clientes_clienteId: createProjetoDto.clientes_clienteId,
+        nome: createProjetoDto.nome
+      }
     });
 
-    return projeto;
+    if(projeto) {
+      throw new HttpException('Já existe um projeto com o mesmo nome para esse cliente', HttpStatus.BAD_REQUEST);
+    }
+
+    // Criando os ambientes padrões
+    if (!createProjetoDto.ambientes) {
+      createProjetoDto.ambientes = {
+        create: [
+          { nome: 'Produção' },
+          { nome: 'Teste' },
+          { nome: 'Desenvolvimento' },
+        ]
+      };
+    }
+
+    return this.prisma.projetos.create({
+      data: createProjetoDto,
+      include: {
+        ambientes: true
+      }
+    });
   }
 
   async findAll() {
-    const projetos = await this.prisma.projetos.findMany({
+    return this.prisma.projetos.findMany({
       include: {
         clientes: {
           select: {
             clienteId: true
           }
-        }
+        },
+        ambientes: true
       }
     });
-
-    return projetos;
   } 
 
   async findOne(id: number) {
@@ -36,21 +58,43 @@ export class ProjetosService {
       },
     });
 
+    if(!projeto) {
+      throw new NotFoundException();
+    }
+
     return projeto;
   }
 
   async update(id: number, updateProjetoDto: UpdateProjetoDto) {
-    const projeto = await this.prisma.projetos.update({
+    const projeto = await this.prisma.projetos.findUnique({
+      where: {
+        projetoId: id
+      }
+    });
+
+    if(!projeto) {
+      throw new NotFoundException();
+    }
+
+    return this.prisma.projetos.update({
       where: {
         projetoId: id,
       },
       data: updateProjetoDto,
     });
-
-    return projeto;
   }
 
   async remove(id: number) {
+    const projeto = await this.prisma.projetos.findUnique({
+      where: {
+        projetoId: id
+      }
+    });
+
+    if(!projeto) {
+      throw new NotFoundException();
+    }
+
     return this.prisma.projetos.delete({
       where: {
         projetoId: id,
